agent/decision_logic/requirements.txt
langchain-core>=0.2.0
langchain-ollama>=0.1.0
langgraph>=0.0.40
pydantic>=2.0.0
python-dotenv>=1.0.0
